Add workflow to run first segment only.
Don't allow value of workflow.resume to change from one run to the next
because that will invalidated hash of each NFCORE_DEMO task.  Instead,
run the first segment in a non-recursive manner to generate a nextflow
log record.  The associated `RUN NAME` or `SESSION ID` can then be used
with resume.  Each segment, including the first, is then run with
'-resume'.